requirements.txt
python-telegram-bot[job-queue]==20.7
selenium==4.25.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
holidays==0.54
nest-asyncio==1.6.0
pytz==2024.1
